Give pathpoint a separate file

+++ roboteam_ai/src/control/positionControllers/ForcePosControl.h
@@ -22,7 +22,7 @@ private:
     const double FORCE_WEIGHT_THEM = 2.0;
     const double FORCE_WEIGHT_BALL = 1.0;
     const double FORCE_WEIGHT_FIELD_SIDES = 1.0;
-    const double POINT_IN_FIELD_MARGIN = 0.5;
+    const float POINT_IN_FIELD_MARGIN = 0.5;
 };
 
 } // control
